+++ app/src/main/java/com/stepstone/apprating/sample/SampleActivity.kt
@@ -4,11 +4,9 @@ import android.os.Bundle
 import android.support.v4.app.FragmentActivity
 import android.widget.Button
 import android.widget.Toast
-
 import com.stepstone.apprating.AppRatingDialog
 import com.stepstone.apprating.listener.RatingDialogListener
-
-import java.util.Arrays
+import java.util.*
 
 class SampleActivity : FragmentActivity(), RatingDialogListener {
 
@@ -38,7 +36,6 @@ class SampleActivity : FragmentActivity(), RatingDialogListener {
                 .setDescriptionTextColor(R.color.descriptionTextColor)
                 .setCommentTextColor(R.color.commentTextColor)
                 .setCommentBackgroundColor(R.color.colorPrimaryDark)
-                //.setPositiveButtonClickedListener(listener)
                 .setWindowAnimation(R.style.MyDialogSlideHorizontalAnimation)
                 .create(this@SampleActivity)
                 .show()